RA-14 fun initPortionsSeekBar(ingredientsAdapter: IngredientsAdapter) in RecipeFragment
