+++ app/src/main/java/com/example/androiddevelopmentapplicationapp/Ingredient.kt
@@ -8,5 +8,4 @@ data class Ingredient(
     val quantity: String,
     val unitOfMeasure: String,
     val description: String,
-    val amount: Float
 ) : Parcelable
