+++ app/src/main/java/com/example/androiddevelopmentapplicationapp/recipes.kt
@@ -48,7 +48,7 @@ object STUB {
             id = 0,
             title = "Классический гамбургер",
             ingredients = listOf(
-                Ingredient("1", "шт.", "Булочка для гамбургера",),
+                Ingredient("1", "шт.","Булочка для гамбургера",),
                 Ingredient("150", "г", "Фарш говяжий",),
                 Ingredient("0.5", "шт.", "Лук репчатый",),
                 Ingredient("1", "лист", "Салат айсберг",),
